requirements.txt
asyncio
google-generativeai
google-adk